requirements.txt
pillow==7.2.0
